Fix use of continuous scales split over facets (#539)
For some reason none of the reference tests showed that there was a
problem with using the non-merged continuous scales to build the
entries. This meant that the scaling in each facet was based on the
local data and not the global data, possibly giving a wrong legend for
some of the facets.