Pregrouped matrices docs (#676)
Closes https://github.com/MakieOrg/AlgebraOfGraphics.jl/issues/267

Clarifies that the passing of 2D data via `pregrouped` must be seen as
experimental, as the rules for this are a bit unclear, yet. Just because
it happens to work for some cases does not mean it was intended to or
will be supported. I have seen some uses in the wild, though, so an
explicit mention seems warranted.